+++ application/language/estonian/translations_lang.php
@@ -446,5 +446,7 @@ $lang['blocked_periods'] = 'Blocked Periods';
 $lang['blocked_period_save'] = 'Blocked Period Save';
 $lang['blocked_period_delete'] = 'Blocked Period Delete';
 $lang['blocked_periods_hint'] = 'Define periods of time where public bookings will be disabled for all providers (e.g. closed dates, holidays etc.).';
-$lang['auxiliary_field'] = 'Auxiliary Field';
+$lang['custom_field'] = 'Custom Field';
+$lang['custom_fields'] = 'Custom Fields';
+$lang['label'] = 'Label';
 // End