+++ application/views/components/appointments_modal.php
@@ -43,27 +43,21 @@
                                         <?php
                                         // Group services by category, only if there is at least one service
                                         // with a parent category.
-                                        $has_category = FALSE;
+                                        $has_category = false;
 
-                                        foreach ($available_services as $service)
-                                        {
-                                            if ( ! empty($service['category_id']))
-                                            {
-                                                $has_category = TRUE;
+                                        foreach ($available_services as $service) {
+                                            if (!empty($service['category_id'])) {
+                                                $has_category = true;
                                                 break;
                                             }
                                         }
 
-                                        if ($has_category)
-                                        {
+                                        if ($has_category) {
                                             $grouped_services = [];
 
-                                            foreach ($available_services as $service)
-                                            {
-                                                if ( ! empty($service['category_id']))
-                                                {
-                                                    if ( ! isset($grouped_services[$service['category_name']]))
-                                                    {
+                                            foreach ($available_services as $service) {
+                                                if (!empty($service['category_id'])) {
+                                                    if (!isset($grouped_services[$service['category_name']])) {
                                                         $grouped_services[$service['category_name']] = [];
                                                     }
 
@@ -75,40 +69,39 @@
                                             // another iteration only for the uncategorized services.
                                             $grouped_services['uncategorized'] = [];
 
-                                            foreach ($available_services as $service)
-                                            {
-                                                if ($service['category_id'] == NULL)
-                                                {
+                                            foreach ($available_services as $service) {
+                                                if ($service['category_id'] == null) {
                                                     $grouped_services['uncategorized'][] = $service;
                                                 }
                                             }
 
-                                            foreach ($grouped_services as $key => $group)
-                                            {
-                                                $group_label = $key !== 'uncategorized'
-                                                    ? e($group[0]['category_name'])
-                                                    : 'Uncategorized';
+                                            foreach ($grouped_services as $key => $group) {
+                                                $group_label =
+                                                    $key !== 'uncategorized'
+                                                        ? e($group[0]['category_name'])
+                                                        : 'Uncategorized';
 
-                                                if (count($group) > 0)
-                                                {
+                                                if (count($group) > 0) {
                                                     echo '<optgroup label="' . $group_label . '">';
 
-                                                    foreach ($group as $service)
-                                                    {
-                                                        echo '<option value="' . $service['id'] . '">'
-                                                            . e($service['name']) . '</option>';
+                                                    foreach ($group as $service) {
+                                                        echo '<option value="' .
+                                                            $service['id'] .
+                                                            '">' .
+                                                            e($service['name']) .
+                                                            '</option>';
                                                     }
 
                                                     echo '</optgroup>';
                                                 }
                                             }
-                                        }
-                                        else
-                                        {
-                                            foreach ($available_services as $service)
-                                            {
-                                                echo '<option value="' . $service['id'] . '">'
-                                                    . e($service['name']) . '</option>';
+                                        } else {
+                                            foreach ($available_services as $service) {
+                                                echo '<option value="' .
+                                                    $service['id'] .
+                                                    '">' .
+                                                    e($service['name']) .
+                                                    '</option>';
                                             }
                                         }
                                         ?>
@@ -124,7 +117,7 @@
                                 </div>
 
                                 <div class="mb-3">
-                                    <?php component('color_selection', ['attributes' => 'id="appointment-color"']) ?>
+                                    <?php component('color_selection', ['attributes' => 'id="appointment-color"']); ?>
                                 </div>
 
                                 <div class="mb-3">
@@ -143,7 +136,7 @@
                                             <option value="<?= e($appointment_status_option) ?>">
                                                 <?= e($appointment_status_option) ?>
                                             </option>
-                                        <?php endforeach ?>
+                                        <?php endforeach; ?>
                                     </select>
                                 </div>
                             </div>
@@ -228,7 +221,7 @@
                                         <?= lang('first_name') ?>
                                         <?php if ($require_first_name): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="first-name"
                                            class="<?= $require_first_name ? 'required' : '' ?> form-control"
@@ -240,7 +233,7 @@
                                         <?= lang('last_name') ?>
                                         <?php if ($require_last_name): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="last-name"
                                            class="<?= $require_last_name ? 'required' : '' ?> form-control"
@@ -252,7 +245,7 @@
                                         <?= lang('email') ?>
                                         <?php if ($require_email): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="email"
                                            class="<?= $require_email ? 'required' : '' ?> form-control"
@@ -264,7 +257,7 @@
                                         <?= lang('phone_number') ?>
                                         <?php if ($require_phone_number): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="phone-number" maxlength="60"
                                            class="<?= $require_phone_number ? 'required' : '' ?> form-control"/>
@@ -280,7 +273,7 @@
                                             <option value="<?= $available_language ?>">
                                                 <?= ucfirst($available_language) ?>
                                             </option>
-                                        <?php endforeach ?>
+                                        <?php endforeach; ?>
                                     </select>
                                 </div>
                             </div>
@@ -290,7 +283,7 @@
                                         <?= lang('address') ?>
                                         <?php if ($require_address): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="address"
                                            class="<?= $require_address ? 'required' : '' ?> form-control"
@@ -302,7 +295,7 @@
                                         <?= lang('city') ?>
                                         <?php if ($require_city): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="city"
                                            class="<?= $require_city ? 'required' : '' ?> form-control"
@@ -314,7 +307,7 @@
                                         <?= lang('zip_code') ?>
                                         <?php if ($require_zip_code): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <input type="text" id="zip-code"
                                            class="<?= $require_zip_code ? 'required' : '' ?> form-control"
@@ -329,7 +322,7 @@
                                     <?php component('timezone_dropdown', [
                                         'attributes' => 'id="timezone" class="form-control required"',
                                         'grouped_timezones' => vars('grouped_timezones')
-                                    ]) ?>
+                                    ]); ?>
                                 </div>
 
                                 <div class="mb-3">
@@ -337,11 +330,15 @@
                                         <?= lang('notes') ?>
                                         <?php if ($require_notes): ?>
                                             <span class="text-danger">*</span>
-                                        <?php endif ?>
+                                        <?php endif; ?>
                                     </label>
                                     <textarea id="customer-notes" rows="2"
                                               class="<?= $require_notes ? 'required' : '' ?> form-control"></textarea>
                                 </div>
+                            </div>
+
+                            <div class="mb-3">
+                                <?php component('custom_fields'); ?>
                             </div>
                         </div>
                     </fieldset>
@@ -361,8 +358,8 @@
     </div>
 </div>
 
-<?php section('scripts') ?>
+<?php section('scripts'); ?>
 
 <script src="<?= asset_url('assets/js/components/appointments_modal.js') ?>"></script>
 
-<?php end_section('scripts') ?>
+<?php end_section('scripts'); ?>