+++ application/views/pages/booking_settings.php
@@ -247,6 +247,50 @@
                         </div>
 
                         <h5 class="text-black-50 mb-3 fw-light">
+                            <?= lang('custom_fields') ?>
+                        </h5>
+
+                        <div class="row mb-5 fields-row">
+                            <?php for ($i = 1; $i <= 5; $i++): ?>
+                                <div class="col-sm-6">
+                                    <div class="form-group mb-5">
+                                        <label for="first-name" class="form-label">
+                                            <?= lang('custom_field') ?> #<?= $i ?>
+                                            <span class="text-danger">*</span>
+                                        </label>
+
+                                        <input type="text" id="custom-field-<?= $i ?>" class="form-control mb-2"
+                                               placeholder="<?= lang('label') ?>"
+                                               data-field="label_custom_field_<?= $i ?>"
+                                               aria-label="label"
+                                        />
+
+                                        <div class="d-flex">
+                                            <div class="form-check form-switch me-4">
+                                                <input class="form-check-input display-switch" type="checkbox"
+                                                       id="display-custom-field-<?= $i ?>"
+                                                       data-field="display_custom_field_<?= $i ?>">
+                                                <label class="form-check-label" for="display-custom-field-<?= $i ?>">
+                                                    <?= lang('display') ?>
+                                                </label>
+                                            </div>
+
+                                            <div class="form-check form-switch">
+                                                <input class="form-check-input require-switch" type="checkbox"
+                                                       id="require-custom-field-<?= $i ?>"
+                                                       data-field="require_custom_field_<?= $i ?>">
+                                                <label class="form-check-label" for="require-custom-field-<?= $i ?>">
+                                                    <?= lang('require') ?>
+                                                </label>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            <?php endfor; ?>
+                        </div>
+
+
+                        <h5 class="text-black-50 mb-3 fw-light">
                             <?= lang('options') ?>
                         </h5>